requirements.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
djangorestframework==3.14.0
idna==3.4
mysqlclient==2.1.1
Pillow==9.5.0
pycodestyle==2.10.0
python-decouple==3.8
pytz==2023.3
requests==2.29.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.15
